Extract delegation strategy to own component
